+++ package.json
@@ -68,7 +68,6 @@
     "react-helmet": "~6.0.0",
     "react-live": "^2.2.2",
     "react-markdown": "^4.3.1",
-    "react-responsive-carousel": "^3.2.7",
     "react-share": "^4.1.0",
     "react-typography": "^0.16.19",
     "rehype-document": "^5.0.0",
